fix: merge from main

+++ cloud-functions/clean-cloud-functions.sh
@@ -16,32 +16,41 @@
 
 PROXY_NAME=cloud-function-http-trigger
 
-delete_apiproxy() {
-  local proxy_name=$1
-  printf "Checking Proxy %s\n" "${proxy_name}"
-  if apigeecli apis get --name "$proxy_name" --org "$PROJECT" --token "$TOKEN" --disable-check >/dev/null 2>&1; then
-    OUTFILE=$(mktemp /tmp/apigee-samples.apigeecli.out.XXXXXX)
-    if apigeecli apis listdeploy --name "$proxy_name" --org "$PROJECT" --token "$TOKEN" --disable-check >"$OUTFILE" 2>&1; then
-      NUM_DEPLOYS=$(jq -r '.deployments | length' "$OUTFILE")
-      if [[ $NUM_DEPLOYS -ne 0 ]]; then
-        echo "Undeploying ${proxy_name}"
-        for ((i = 0; i < NUM_DEPLOYS; i++)); do
-          ENVNAME=$(jq -r ".deployments[$i].environment" "$OUTFILE")
-          REV=$(jq -r ".deployments[$i].revision" "$OUTFILE")
-          apigeecli apis undeploy --name "${proxy_name}" --env "$ENVNAME" --rev "$REV" --org "$PROJECT" --token "$TOKEN" --disable-check
-        done
-      else
-        printf "  There are no deployments of %s to remove.\n" "${proxy_name}"
-      fi
+delete_service_account() {
+    local sa_email="$1" project="$2"
+    if gcloud iam service-accounts describe "$sa_email" --project="$project" >/dev/null 2>&1; then
+        gcloud iam service-accounts delete "$sa_email" --project="$project" --quiet
+    else
+        printf "  That service account does not exist.\n"
     fi
-    [[ -f "$OUTFILE" ]] && rm "$OUTFILE"
+}
 
-    echo "Deleting proxy ${proxy_name}"
-    apigeecli apis delete --name "${proxy_name}" --org "$PROJECT" --token "$TOKEN" --disable-check
+delete_apiproxy() {
+    local proxy_name=$1
+    printf "Checking Proxy %s\n" "${proxy_name}"
+    if apigeecli apis get --name "$proxy_name" --org "$APIGEE_PROJECT" --token "$TOKEN" --disable-check >/dev/null 2>&1; then
+        OUTFILE=$(mktemp /tmp/apigee-samples.apigeecli.out.XXXXXX)
+        if apigeecli apis listdeploy --name "$proxy_name" --org "$APIGEE_PROJECT" --token "$TOKEN" --disable-check >"$OUTFILE" 2>&1; then
+            NUM_DEPLOYS=$(jq -r '.deployments | length' "$OUTFILE")
+            if [[ $NUM_DEPLOYS -ne 0 ]]; then
+                echo "Undeploying ${proxy_name}"
+                for ((i = 0; i < NUM_DEPLOYS; i++)); do
+                    ENVNAME=$(jq -r ".deployments[$i].environment" "$OUTFILE")
+                    REV=$(jq -r ".deployments[$i].revision" "$OUTFILE")
+                    apigeecli apis undeploy --name "${proxy_name}" --env "$ENVNAME" --rev "$REV" --org "$APIGEE_PROJECT" --token "$TOKEN" --disable-check
+                done
+            else
+                printf "  There are no deployments of %s to remove.\n" "${proxy_name}"
+            fi
+        fi
+        [[ -f "$OUTFILE" ]] && rm "$OUTFILE"
 
-  else
-    printf "  The proxy %s does not exist.\n" "${proxy_name}"
-  fi
+        echo "Deleting proxy ${proxy_name}"
+        apigeecli apis delete --name "${proxy_name}" --org "$APIGEE_PROJECT" --token "$TOKEN" --disable-check
+
+    else
+        printf "  The proxy %s does not exist.\n" "${proxy_name}"
+    fi
 }
 
 MISSING_ENV_VARS=()
@@ -53,9 +62,9 @@ MISSING_ENV_VARS=()
 [[ -z "$CLOUD_FUNCTION_NAME" ]] && MISSING_ENV_VARS+=('CLOUD_FUNCTION_NAME')
 
 [[ ${#MISSING_ENV_VARS[@]} -ne 0 ]] && {
-  printf -v joined '%s,' "${MISSING_ENV_VARS[@]}"
-  printf "You must set these environment variables: %s\n" "${joined%,}"
-  exit 1
+    printf -v joined '%s,' "${MISSING_ENV_VARS[@]}"
+    printf "You must set these environment variables: %s\n" "${joined%,}"
+    exit 1
 }
 
 TOKEN=$(gcloud auth print-access-token)
@@ -66,24 +75,24 @@ export PATH=$PATH:$HOME/.apigeecli/bin
 
 delete_apiproxy "${PROXY_NAME}"
 
-printf "Deleting the proxy service account\n"
+printf "Deleting the proxy service account...\n"
 PROXY_SA_NAME=proxy-apigee-sample-sa-1
 PROXY_SA_EMAIL="${PROXY_SA_NAME}@${APIGEE_PROJECT}.iam.gserviceaccount.com"
-gcloud iam service-accounts delete "${PROXY_SA_EMAIL}" \
-  --project="${APIGEE_PROJECT}" \
-  --quiet
+delete_service_account "$PROXY_SA_EMAIL" "$APIGEE_PROJECT"
 
-printf "Deleting the cloud function\n"
-gcloud functions delete "$CLOUD_FUNCTION_NAME" \
-  --region="$CLOUD_FUNCTIONS_REGION" \
-  --project="$CLOUD_FUNCTIONS_PROJECT" \
-  --quiet
+printf "Deleting the cloud function...\n"
+if gcloud functions describe "$CLOUD_FUNCTION_NAME" --region="$CLOUD_FUNCTIONS_REGION" --project="$CLOUD_FUNCTIONS_PROJECT" >/dev/null 2>&1; then
+    gcloud functions delete "$CLOUD_FUNCTION_NAME" \
+        --region="$CLOUD_FUNCTIONS_REGION" \
+        --project="$CLOUD_FUNCTIONS_PROJECT" \
+        --quiet
+else
+    printf "  That Cloud function does not exist.\n"
+fi
 
-printf "Deleting the cloud function service account\n"
+printf "Deleting the cloud function service account...\n"
 CF_SA_NAME=cf-apigee-sample-sa-1
 CF_SA_EMAIL="${CF_SA_NAME}@${CLOUD_FUNCTIONS_PROJECT}.iam.gserviceaccount.com"
-gcloud iam service-accounts delete "${CF_SA_EMAIL}" \
-  --project="$CLOUD_FUNCTIONS_PROJECT" \
-  --quiet
+delete_service_account "${CF_SA_EMAIL}" "$CLOUD_FUNCTIONS_PROJECT"
 
 printf "\nAll the artifacts for this sample have been removed.\n"
